requirements.txt
asterisk_ami==0.1.6
requests==2.28.2
